Move CTransaction and related classes to separate module
